fix warning "implicit conversion loses floating-point precision"

+++ src/Aruco/ArucoCore.cpp
@@ -49,10 +49,10 @@ const QMatrix4x4 ArucoCore::getDetectedMatrix( int markerId, cv::Mat inputImage 
 	//this->detectMarkers();
 	this->getMatrix( markerId, modelViewMatrix );
 
-	QMatrix4x4 matrix( modelViewMatrix[ 0], modelViewMatrix[ 1], modelViewMatrix[ 2], modelViewMatrix[ 3],
-					   modelViewMatrix[ 4], modelViewMatrix[ 5], modelViewMatrix[ 6], modelViewMatrix[ 7],
-					   modelViewMatrix[ 8], modelViewMatrix[ 9], modelViewMatrix[10], modelViewMatrix[11],
-					   modelViewMatrix[12], modelViewMatrix[13], modelViewMatrix[14], modelViewMatrix[15] );
+	QMatrix4x4 matrix( static_cast<float>( modelViewMatrix[ 0] ), static_cast<float>( modelViewMatrix[ 1] ), static_cast<float>( modelViewMatrix[ 2] ), static_cast<float>( modelViewMatrix[ 3] ),
+					   static_cast<float>( modelViewMatrix[ 4] ), static_cast<float>( modelViewMatrix[ 5] ), static_cast<float>( modelViewMatrix[ 6] ), static_cast<float>( modelViewMatrix[ 7] ),
+					   static_cast<float>( modelViewMatrix[ 8] ), static_cast<float>( modelViewMatrix[ 9] ), static_cast<float>( modelViewMatrix[10] ), static_cast<float>( modelViewMatrix[11] ),
+					   static_cast<float>( modelViewMatrix[12] ), static_cast<float>( modelViewMatrix[13] ), static_cast<float>( modelViewMatrix[14] ), static_cast<float>( modelViewMatrix[15] ) );
 
 	return matrix;
 }
@@ -66,10 +66,10 @@ const QMatrix4x4 ArucoCore::getProjectionMatrix( cv::Mat inputImage )
 	//get projection matrix via ArUco
 	mCamParam.glGetProjectionMatrix( inputImage.size(),inputImage.size(),projectionMatrix,0.01,10000.0 );
 
-	QMatrix4x4 matrix( projectionMatrix[ 0], projectionMatrix[ 1], projectionMatrix[ 2], projectionMatrix[ 3],
-					   projectionMatrix[ 4], projectionMatrix[ 5], projectionMatrix[ 6], projectionMatrix[ 7],
-					   projectionMatrix[ 8], projectionMatrix[ 9], projectionMatrix[10], projectionMatrix[11],
-					   projectionMatrix[12], projectionMatrix[13], projectionMatrix[14], projectionMatrix[15] );
+	QMatrix4x4 matrix( static_cast<float>( projectionMatrix[ 0] ), static_cast<float>( projectionMatrix[ 1] ), static_cast<float>( projectionMatrix[ 2] ), static_cast<float>( projectionMatrix[ 3] ),
+					   static_cast<float>( projectionMatrix[ 4] ), static_cast<float>( projectionMatrix[ 5] ), static_cast<float>( projectionMatrix[ 6] ), static_cast<float>( projectionMatrix[ 7] ),
+					   static_cast<float>( projectionMatrix[ 8] ), static_cast<float>( projectionMatrix[ 9] ), static_cast<float>( projectionMatrix[10] ), static_cast<float>( projectionMatrix[11] ),
+					   static_cast<float>( projectionMatrix[12] ), static_cast<float>( projectionMatrix[13] ), static_cast<float>( projectionMatrix[14] ), static_cast<float>( projectionMatrix[15] ) );
 	return matrix;
 }
