+++ src/Leap/HandModule/Manipulator/HandObjectManipulator.cpp
@@ -42,7 +42,7 @@ void Leap::HandObjectManipulator::updateHands( Leap::Hand leftHand, Leap::Hand r
 
 	// update lavej ruky
 	if ( leftHand.isValid() ) {
-		Leap::Vector lVector = Leap::Vector( this->center[0]+0.5,this->center[1],this->center[2] );
+		Leap::Vector lVector = Leap::Vector( static_cast<float>( this->center[0] )+0.5f,static_cast<float>( this->center[1] ),static_cast<float>( this->center[2] ) );
 		//ziskanie pozicie dlane
 		lVector = leftHand.palmPosition();
 
@@ -64,11 +64,11 @@ void Leap::HandObjectManipulator::updateHands( Leap::Hand leftHand, Leap::Hand r
 		osg::Matrix view = camera->getViewMatrix();
 		osg::Matrix proj = camera->getProjectionMatrix();
 		osg::Matrix model = leftPalm->getWorldMatrices()[0];
-		osg::Vec3 world_coords = osg::Vec3( static_cast<double>( this->center[0] ) + this->direction[0] + static_cast<double>( lVector.x )/100.0,
-											static_cast<double>( this->center[1] )+this->direction[1] +static_cast<double>( lVector.y )/100.0,
-											static_cast<double>( this->center[2] )+this->direction[2] +static_cast<double>( lVector.z )/100.0 );
+		osg::Vec3 world_coords = osg::Vec3( static_cast<float>( this->center[0] ) + static_cast<float>( this->direction[0] ) + lVector.x/100.0f,
+											static_cast<float>( this->center[1] ) + static_cast<float>( this->direction[1] ) + lVector.y/100.0f,
+											static_cast<float>( this->center[2] ) + static_cast<float>( this->direction[2] ) + lVector.z/100.0f );
 		osg::Vec3 screenCoords = world_coords * view * proj * win;
-		screenCoords.set( ( ( static_cast<double>( screenCoords.x() ) / viewport->width() ) * 640 ), ( static_cast<double>( screenCoords.y() ) / viewport->height() ) * 480, screenCoords.z() );
+		screenCoords.set( ( static_cast<float>( screenCoords.x() ) / static_cast<float>( viewport->width() ) ) * 640, ( static_cast<float>( screenCoords.y() ) / static_cast<float>( viewport->height() ) ) * 480, screenCoords.z() );
 		this->mapper->setNodeScreenCoords( screenCoords );
 
 		// update prstov lavej ruky
@@ -79,7 +79,7 @@ void Leap::HandObjectManipulator::updateHands( Leap::Hand leftHand, Leap::Hand r
 
 	// update pravej ruky
 	if ( rightHand.isValid() ) {
-		Leap::Vector rVector = Leap::Vector( this->center[0]-0.5,this->center[1],this->center[2] );
+		Leap::Vector rVector = Leap::Vector( static_cast<float>( this->center[0] )-0.5f,static_cast<float>( this->center[1] ),static_cast<float>( this->center[2] ) );
 		//ziskanie pozicie dlane
 		rVector = rightHand.palmPosition();
